+++ spec/rails/templates/erb_integration_system_spec/app/controllers/erb_integration_system_spec_controller.rb
@@ -2,6 +2,12 @@ class ErbIntegrationSystemSpecController < ApplicationController
   def erb_to_widget_with_render_partial
   end
 
+  def erb_to_widget_with_widget
+  end
+
+  def erb_to_widget_with_widget_class
+  end
+
   def prefers_erb_partial
   end