+++ spec/rails/templates/erb_integration_system_spec/app/views/erb_integration_system_spec/erb_to_widget_with_widget_widget.rb
@@ -1,0 +1,7 @@
+class Views::ErbIntegrationSystemSpec::ErbToWidgetWithWidgetWidget < Fortitude::Widgets::Html5
+  needs :name
+
+  def content
+    p "erb to widget with widget widget, name #{name}"
+  end
+end