+++ spec/rails/templates/erector_coexistence_system_spec/app/views/erector_coexistence_system_spec/erector_widget.rb
@@ -1,0 +1,7 @@
+class Views::ErectorCoexistenceSystemSpec::ErectorWidget < ::Erector::Widget
+  needs :name
+
+  def content
+    p "this is an Erector widget, #{@name}"
+  end
+end