+++ spec/rails/templates/erector_coexistence_system_spec/app/views/erector_coexistence_system_spec/fortitude_widget.rb
@@ -1,0 +1,7 @@
+class Views::ErectorCoexistenceSystemSpec::FortitudeWidget < ::Fortitude::Widgets::Html5
+  needs :name
+
+  def content
+    p "this is a Fortitude widget, #{name}"
+  end
+end